refactor: move map components into map module
