+++ src/App/App.js
@@ -1,23 +1,17 @@
 import "@hyperobjekt/mapgl/dist/style.css";
 import "@hyperobjekt/scales/dist/style.css";
-import { Legend } from "./Legend/Legend";
-import { Button, CssBaseline } from "@mui/material";
+import { CssBaseline } from "@mui/material";
 import { styled, ThemeProvider } from "@mui/system";
 import Header from "./components/Header";
 import { IndicatorPanel } from "./IndicatorPanel/IndicatorPanel";
 import useIndicatorPanelStore from "./IndicatorPanel/store";
 import shallow from "zustand/shallow";
-import { MapTooltip, Map } from "./Map";
+import { Map } from "./Map";
 import CustomizeIndiactorPanel from "./IndicatorPanel/CustomizeIndicatorPanel";
 import Dashboard, { QueryParamRouter } from "@hyperobjekt/react-dashboard";
 import { Scorecards } from "./Scorecards";
 import theme from "../theme";
 import SearchModal from "./Search/components/SearchModal";
-
-// // debug tools
-// import Debug from "./Demo/components/Debug";
-// import { ReactQueryDevtools } from "react-query/devtools";
-// const DEBUG = ["context", "mapSources", "metrics"];
 
 const CONFIG = {
   app: "/assets/config/app.json",
@@ -44,23 +38,6 @@ const AppWrapper = styled("div")({
   overflow: "hidden",
 });
 
-const MapBodyWrapper = styled("div")(({ theme }) => ({
-  position: "relative",
-  display: "flex",
-  height: `calc(100vh - ${theme.spacing(7)})`,
-  minHeight: `calc(100vh - ${theme.spacing(7)})`,
-  overflow: "hidden",
-  [theme.breakpoints.up("sm")]: {
-    height: `calc(100vh - ${theme.spacing(8)})`,
-    minHeight: `calc(100vh - ${theme.spacing(8)})`,
-  },
-  "& .HypMapGl-root": {
-    position: "absolute",
-    inset: 0,
-    height: "100%",
-  },
-}));
-
 function App() {
   // pull state and setter that determines if the indicator panel is open
   const [indicatorsOpen, setIndicatorsOpen] = useIndicatorPanelStore(
@@ -77,29 +54,13 @@ function App() {
         <QueryParamRouter />
         <AppWrapper className="App">
           <Header />
-          <MapBodyWrapper>
-            <Map>
-              <Legend square>
-                {!indicatorsOpen && (
-                  <Button
-                    variant="contained"
-                    size="small"
-                    fullWidth
-                    sx={{ mt: 1 }}
-                    onClick={() => setIndicatorsOpen(!indicatorsOpen)}
-                  >
-                    Show all indicators
-                  </Button>
-                )}
-              </Legend>
-            </Map>
+          <Map>
             <IndicatorPanel
               open={indicatorsOpen}
               onClose={() => setIndicatorsOpen(false)}
             />
             <CustomizeIndiactorPanel open={customizeOpen} />
-            <MapTooltip />
-          </MapBodyWrapper>
+          </Map>
           <Scorecards />
         </AppWrapper>
         <SearchModal />
