+++ src/App/Legend/index.js
@@ -1,0 +1,1 @@
+export { default as Legend } from "./Legend";
